+++ @navikt/core/react/src/util/focus-boundary/FocusBoundary.tsx
@@ -6,18 +6,10 @@ import React, {
   useState,
 } from "react";
 import { Slot } from "../../slot/Slot";
-import { useMergeRefs } from "../../util/hooks";
-import { useEventCallback } from "../hooks/useEventCallback";
-
-const AUTOFOCUS_ON_MOUNT = "focusBoundary.autoFocusOnMount";
-const AUTOFOCUS_ON_UNMOUNT = "focusBoundary.autoFocusOnUnmount";
-const EVENT_OPTIONS = { bubbles: false, cancelable: true };
-
-/* TODO: Regular story */
-/**
- * TODO:
- * - owner(container) for corrent document
- */
+import { useClientLayoutEffect, useMergeRefs } from "../../util/hooks";
+import { hideNonTargetElements } from "../hideNonTargetElements";
+import { useLatestRef } from "../hooks/useLatestRef";
+import { ownerDocument } from "../owner";
 
 /* -------------------------------------------------------------------------- */
 /*                                 FocusBoundary                                 */
@@ -48,15 +40,31 @@ interface FocusBoundaryProps extends React.HTMLAttributes<HTMLDivElement> {
    */
   trapped?: boolean;
   /**
-   * Event handler called when auto-focusing on mount.
-   * Can be prevented.
+   * Will try to focus the given element on mount.
+   *
+   * If not provided, FocusBoundary will try to focus the first
+   * tabbable element inside the boundary.
+   *
+   * Set to `false` to not focus anything.
    */
-  onMountAutoFocus?: (event: Event) => void;
+  initialFocus?:
+    | boolean
+    | React.MutableRefObject<HTMLElement | null>
+    | (() => boolean | HTMLElement | null | undefined);
   /**
-   * Event handler called when auto-focusing on unmount.
-   * Can be prevented.
+   * Will try to focus the given element on unmount.
+   *
+   * If not provided, FocusBoundary will try to focus the element
+   * that was focused before the FocusBoundary mounted.
+   *
+   * Set to `false` to not focus anything.
    */
-  onUnmountAutoFocus?: (event: Event) => void;
+  returnFocus?: boolean | React.MutableRefObject<HTMLElement | null>;
+  /**
+   * Hides all outside content from screen readers when true.
+   * @default false
+   */
+  modal?: boolean;
 }
 
 const FocusBoundary = forwardRef<HTMLDivElement, FocusBoundaryProps>(
@@ -64,14 +72,15 @@ const FocusBoundary = forwardRef<HTMLDivElement, FocusBoundaryProps>(
     {
       loop = false,
       trapped = false,
-      onMountAutoFocus: onMountAutoFocusProp,
-      onUnmountAutoFocus: onUnmountAutoFocusProp,
+      initialFocus = true,
+      returnFocus = true,
+      modal = false,
       ...restProps
     }: FocusBoundaryProps,
     forwardedRef,
   ) => {
-    const onMountAutoFocus = useEventCallback(onMountAutoFocusProp);
-    const onUnmountAutoFocus = useEventCallback(onUnmountAutoFocusProp);
+    const initialFocusRef = useLatestRef(initialFocus);
+    const returnFocusRef = useLatestRef(returnFocus);
 
     const lastFocusedElementRef = useRef<HTMLElement | null>(null);
     const [container, setContainer] = useState<HTMLElement | null>(null);
@@ -102,10 +111,7 @@ const FocusBoundary = forwardRef<HTMLDivElement, FocusBoundaryProps>(
         if (container.contains(target)) {
           lastFocusedElementRef.current = target;
         } else {
-          queueFocus(lastFocusedElementRef.current, {
-            select: true,
-            sync: true,
-          });
+          focus(lastFocusedElementRef.current, { select: true });
         }
       }
 
@@ -138,10 +144,7 @@ const FocusBoundary = forwardRef<HTMLDivElement, FocusBoundaryProps>(
          * when they are not supposed to (like when clicking on elements outside the container
          */
         if (!container.contains(relatedTarget)) {
-          queueFocus(lastFocusedElementRef.current, {
-            select: true,
-            sync: true,
-          });
+          focus(lastFocusedElementRef.current, { select: true });
         }
       }
 
@@ -156,7 +159,7 @@ const FocusBoundary = forwardRef<HTMLDivElement, FocusBoundaryProps>(
         }
 
         if (mutations.some((mutation) => mutation.removedNodes.length > 0)) {
-          queueFocus(container, { sync: true });
+          focus(container);
         }
       };
 
@@ -172,89 +175,127 @@ const FocusBoundary = forwardRef<HTMLDivElement, FocusBoundaryProps>(
       };
     }, [trapped, container, focusBoundary.paused]);
 
-    /* Handles autofocus on mount and unmount */
+    /* Adds element to focus-stack */
     useEffect(() => {
       if (!container) {
         return;
       }
 
       focusBoundarysStack.add(focusBoundary);
-      const initialFocusedElement =
-        document.activeElement as HTMLElement | null;
-      const containsActiveElement =
-        initialFocusedElement && container.contains(initialFocusedElement);
-
-      /*
-       * We only autofocus on mount if container does not contain active element.
-       * If container has an element with `autoFocus` attribute, browser will
-       * have already moved focus there before this effect runs.
-       */
-      if (!containsActiveElement) {
-        const mountEvent = new CustomEvent(AUTOFOCUS_ON_MOUNT, EVENT_OPTIONS);
-        container.addEventListener(AUTOFOCUS_ON_MOUNT, onMountAutoFocus);
-        container.dispatchEvent(mountEvent);
-
-        /* If consumer does not manually prevent event and handle focus themselves */
-        if (!mountEvent.defaultPrevented) {
-          /**
-           * Attempts focusing the first element in a list of candidates.
-           * Stops when focus has actually moved.
-           */
-          const candidates = removeLinks(getTabbableCandidates(container));
-          const previouslyFocusedElement = document.activeElement;
-          for (const candidate of candidates) {
-            queueFocus(candidate, { select: true });
-            if (document.activeElement !== previouslyFocusedElement) {
-              break;
-            }
-          }
-
-          /* focusFirst might not find any candidates, so we fall back to focusing container */
-          if (document.activeElement === initialFocusedElement) {
-            queueFocus(container);
-          }
-        }
-      }
 
       return () => {
-        container.removeEventListener(AUTOFOCUS_ON_MOUNT, onMountAutoFocus);
-
-        /**
-         * https://github.com/facebook/react/issues/17894
-         * We delay to next tick to avoid issues with React's event system
-         * where calling `focus` inside a effect cleanup causes React to not call onFocus handlers.
-         */
         setTimeout(() => {
-          const unmountEvent = new CustomEvent(
-            AUTOFOCUS_ON_UNMOUNT,
-            EVENT_OPTIONS,
-          );
-          container.addEventListener(AUTOFOCUS_ON_UNMOUNT, onUnmountAutoFocus);
-          container.dispatchEvent(unmountEvent);
-
-          /* If consumer does not manually prevent event and handle focus themselves */
-          if (!unmountEvent.defaultPrevented) {
-            /* To avoid CPU-spikes on Chrome, we make sure element is still connected to the DOM. */
-            queueFocus(
-              initialFocusedElement?.isConnected
-                ? initialFocusedElement
-                : document.body,
-              {
-                select: true,
-                sync: true,
-              },
-            );
-          }
-          /* Since this is inside a cleanup, we need to instantly remove the listener ourselves */
-          container.removeEventListener(
-            AUTOFOCUS_ON_UNMOUNT,
-            onUnmountAutoFocus,
-          );
-
           focusBoundarysStack.remove(focusBoundary);
         }, 0);
       };
-    }, [container, onMountAutoFocus, onUnmountAutoFocus, focusBoundary]);
+    }, [container, focusBoundary]);
+
+    useEffect(() => {
+      if (!container || !modal) {
+        return;
+      }
+
+      return hideNonTargetElements([container]);
+    }, [container, modal]);
+
+    /* Handles mount focus */
+    useClientLayoutEffect(() => {
+      if (!container || initialFocusRef.current === false) {
+        return;
+      }
+
+      const ownerDoc = ownerDocument(container);
+      const previouslyFocusedElement = ownerDoc.activeElement;
+
+      queueMicrotask(() => {
+        const focusableElements = removeLinks(getTabbableCandidates(container));
+        const initialFocusValueOrFn = initialFocusRef.current;
+        const resolvedInitialFocus =
+          typeof initialFocusValueOrFn === "function"
+            ? initialFocusValueOrFn()
+            : initialFocusValueOrFn;
+
+        if (
+          resolvedInitialFocus === undefined ||
+          resolvedInitialFocus === false
+        ) {
+          return;
+        }
+
+        let elToFocus: HTMLElement | null | undefined;
+        const fallbackelements = focusableElements[0] || container;
+
+        /* `null` should fallback to default behavior in case of an empty ref. */
+        if (resolvedInitialFocus === true || resolvedInitialFocus === null) {
+          elToFocus = fallbackelements;
+        } else {
+          elToFocus = resolveRef(resolvedInitialFocus) || fallbackelements;
+        }
+
+        const focusAlreadyInsideFloatingEl = container.contains(
+          previouslyFocusedElement,
+        );
+
+        if (focusAlreadyInsideFloatingEl) {
+          return;
+        }
+
+        focus(elToFocus, {
+          preventScroll: elToFocus === container,
+          sync: false,
+        });
+      });
+    }, [container, initialFocusRef]);
+
+    /* Handles unmount focus */
+    useClientLayoutEffect(() => {
+      if (!container) {
+        return;
+      }
+      const ownerDoc = ownerDocument(container);
+      const previouslyFocusedElement = ownerDoc.activeElement;
+
+      function getReturnElement() {
+        let resolvedReturnFocusValue = returnFocusRef.current;
+
+        if (
+          resolvedReturnFocusValue === undefined ||
+          resolvedReturnFocusValue === false
+        ) {
+          return null;
+        }
+
+        /* `null` should fallback to default behavior in case of an empty ref. */
+        if (resolvedReturnFocusValue === null) {
+          resolvedReturnFocusValue = true;
+        }
+
+        if (typeof resolvedReturnFocusValue === "boolean") {
+          const el = previouslyFocusedElement;
+          return el?.isConnected ? el : ownerDoc.body;
+        }
+
+        const fallback = previouslyFocusedElement || ownerDoc.body;
+
+        return resolveRef(resolvedReturnFocusValue) || fallback;
+      }
+
+      return () => {
+        const returnElement = getReturnElement() as HTMLElement | null;
+        const activeEl = ownerDoc.activeElement;
+
+        queueMicrotask(() => {
+          if (
+            // eslint-disable-next-line react-hooks/exhaustive-deps
+            returnFocusRef.current &&
+            returnElement &&
+            returnElement !== activeEl
+          ) {
+            returnElement.focus({ preventScroll: true });
+          }
+        });
+      };
+    }, [container, returnFocusRef]);
 
     /* Takes care of looping focus */
     const handleKeyDown = useCallback(
@@ -294,12 +335,12 @@ const FocusBoundary = forwardRef<HTMLDivElement, FocusBoundaryProps>(
           if (!event.shiftKey && focusedElement === last) {
             event.preventDefault();
             if (loop) {
-              queueFocus(first, { select: true, sync: true });
+              focus(first, { select: true });
             }
           } else if (event.shiftKey && focusedElement === first) {
             event.preventDefault();
             if (loop) {
-              queueFocus(last, { select: true, sync: true });
+              focus(last, { select: true });
             }
           }
         }
@@ -395,40 +436,36 @@ function isHidden(node: HTMLElement, { upTo }: { upTo?: HTMLElement }) {
 }
 
 let rafId = 0;
-
-function queueFocus(
+function focus(
   element?: HTMLElement | null,
-  { select = false, sync = false } = {},
+  { select = false, preventScroll = true, sync = true } = {},
 ) {
   if (!element?.focus) {
     return;
   }
 
-  cancelAnimationFrame(rafId);
   const previouslyFocusedElement = document.activeElement;
 
-  const exec = () => {
-    element?.focus({ preventScroll: true });
-
-    if (!select) {
-      return;
-    }
-
-    /* By default, inputs that gets focus should select its contents */
-    if (
-      element !== previouslyFocusedElement &&
-      element instanceof HTMLInputElement &&
-      "select" in element
-    ) {
-      element.select();
-    }
-  };
+  cancelAnimationFrame(rafId);
+  const exec = () => element.focus({ preventScroll });
 
   if (sync) {
     exec();
   } else {
     rafId = requestAnimationFrame(exec);
   }
+
+  if (!select) {
+    return;
+  }
+
+  /* By default, inputs that gets focus should select its contents */
+  if (
+    element !== previouslyFocusedElement &&
+    element instanceof HTMLInputElement &&
+    "select" in element
+  )
+    element.select();
 }
 
 /* ---------------------------- FocusBoundary stack ---------------------------- */
@@ -469,6 +506,18 @@ function arrayRemove<T>(array: T[], item: T) {
 
 function removeLinks(items: HTMLElement[]) {
   return items.filter((item) => item.tagName !== "A");
+}
+
+/**
+ * If the provided argument is a ref object, returns its `current` value.
+ * Otherwise, returns the argument itself.
+ *
+ * Non-generic to safely handle refs whose `.current` may be `null`.
+ */
+function resolveRef(
+  maybeRef: HTMLElement | React.RefObject<HTMLElement | null | undefined>,
+): HTMLElement | null | undefined {
+  return "current" in maybeRef ? maybeRef.current : maybeRef;
 }
 
 export { FocusBoundary };