+++ aksel.nav.no/website/pages/eksempler/textarea/description.tsx
@@ -20,4 +20,5 @@ export const Demo = {
 
 export const args = {
   index: 2,
+  desc: "Bruk `description` i tillegg til `label` når det er behov for mer utfyllende forklaring.",
 };